requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
boto3==1.18.55
botocore==1.21.55
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coverage==6.0.2
docker==5.0.2
Faker==8.16.0
faker-vehicle==0.2.0
idna==3.2
jmespath==0.10.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
websocket-client==1.2.1
